Rename controller files.
